refactor(console): split the workflows view into canvas, run panels and helpers (#303)
WorkflowsView.tsx was 1872 lines and about to grow an index and a copilot.
Move the pure canvas arithmetic and run-state folds to workflows/graph.ts, the
at-a-glance run reading to workflows/run-health.ts (the cards need the same
reading, and two implementations of "is this run healthy?" would drift), and
the three drawers to their own files. No behaviour change.
